requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
fastmcp>=0.1.0
mcp>=0.1.0
pymongo>=4.0.0
streamlit>=1.30.0
streamlit-calendar>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0